# agent-api/openapi.yaml
openapi: 3.1.0
info:
  title: Metamorphosis Ivette Private Lab Agent API
  version: 0.2.0
  description: Private agent API contract for Ivette's wellness second brain.
servers:
  - url: http://localhost:3000
paths:
...
  /api/brain:
    get:
      summary: Search brain memory
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results
    post:
      summary: Save brain memory entry
      responses:
        '201':
          description: Stored brain memory entry
...